# Conflicts:
#	src/main/java/ru/yandex/practicum/filmorate/service/DatabaseUserService.java

+++ src/main/java/ru/yandex/practicum/filmorate/service/DatabaseUserService.java
@@ -23,6 +23,11 @@ public class DatabaseUserService implements UserService {
 
     private final UserRepository userRepository;
 
+    private boolean checkFriendship(int userId, int friendId) {
+        List<Integer> friendsList = userRepository.getFriends(userId);
+        return friendsList != null && friendsList.contains(friendId);
+    }
+
     @Override
     @Transactional
     public User createUser(User user) {
@@ -48,17 +53,17 @@ public class DatabaseUserService implements UserService {
     }
 
     @Override
-    public List<User> getCommonFriends(int userId, int otherUserId) {
+    public List<Integer> getCommonFriends(int userId, int otherUserId) {
         if (userRepository.findById(userId).isEmpty() ||
                 userRepository.findById(otherUserId).isEmpty()) {
             throw new NotFoundException("Невозможно получить общих друзей: один из пользователей не найден");
         }
 
-        List<User> userFriends = userRepository.getFriends(userId);
-        List<User> otherFriends = userRepository.getFriends(otherUserId);
+        List<Integer> userFriends = userRepository.getFriends(userId);
+        List<Integer> otherFriends = userRepository.getFriends(otherUserId);
 
-        Set<User> otherSet = new HashSet<>(otherFriends);
-        List<User> common = userFriends.stream()
+        Set<Integer> otherSet = new HashSet<>(otherFriends);
+        List<Integer> common = userFriends.stream()
                 .filter(otherSet::contains)
                 .collect(Collectors.toList());
 
@@ -80,11 +85,12 @@ public class DatabaseUserService implements UserService {
     }
 
     @Override
-    public List<User> getUserFriends(int id) {
-        if (userRepository.findById(id).isEmpty()) {
-            throw new NotFoundException("Пользователь с ID " + id + " не найден.");
+    public List<Integer> getUserFriends(int id) {
+        List<Integer> userFriends = userRepository.getFriends(id);
+        if (userFriends.isEmpty()) {
+            throw new NotFoundException("У пользователя ID " + id + " нет друзей.");
         }
-        return userRepository.getFriends(id);
+        return userFriends;
     }
 
     @Override
@@ -99,7 +105,7 @@ public class DatabaseUserService implements UserService {
         if (!userRepository.existsById(userId) || !userRepository.existsById(friendId)) {
             throw new NotFoundException("Невозможно добавить в друзья: один из пользователей не найден.");
         }
-        if (userRepository.existsFriendship(userId, friendId)) {
+        if (checkFriendship(userId, friendId)) {
             throw new OperationNotAllowedException("Пользователи уже являются друзьями.");
         }
         userRepository.makeFriends(userId, friendId);
@@ -116,6 +122,9 @@ public class DatabaseUserService implements UserService {
         }
         if (!userRepository.existsById(userId) || !userRepository.existsById(friendId)) {
             throw new NotFoundException("Невозможно удалить дружбу: один из пользователей не найден.");
+        }
+        if (!checkFriendship(userId, friendId)) {
+            throw new OperationNotAllowedException("Невозможно удалить дружбу: пользователи не являются друзьями.");
         }
         userRepository.removeFriends(userId, friendId);
     }
